7576 -  file naming, refactor html under imports, package.json exports definition for html
