+++ packages/integrations/markdoc/src/extensions/html/htmlattributes/mutate-raw-html-attributes-to-react-compatible-variants.ts
