+++ packages/integrations/markdoc/src/extensions/html/index.ts
@@ -1,0 +1,2 @@
+export { htmlTokenTransform } from "./transform/html-token-transform";
+export { htmlTag } from "./tagdefs/html.tag";
